[android] LocalSearchEngine -> SearchMediator (like on desktop, less confusing name)
Some TorrentSearchResult classes that hadn't been implemented for Android

+++ android/src/com/frostwire/frostclick/TorrentPromotionSearchResult.java
@@ -45,7 +45,7 @@ public class TorrentPromotionSearchResult implements TorrentSearchResult {
     }
 
     @Override
-    public double getSize() {
+    public long getSize() {
         return slide.size;
     }
